Added a workaround for the bug in Hystrix UnitTest code.
`HystrixPlugins.UnitTest.reset()` method is used to reset hystrix internal state between unit tests. In the current version, however, it fails to reset the commandExecutionHook which causes tests that set it to fail if run in sequence.
The workaround is to use reflection to access the private field to reset its state. To that end the `TenacityTestRule` has been updated to exactly that.